added files from React Demo
Added the stuff from the talk I gave in January to the DC user group. The talk was on migrating from Angular to React. I am not sure if dragging and dropping into the UI will leave a project structure that will run properly, as it was done from a tablet. This is more for somebody to take a look at the style of coding I employ. That and how I structure things etc....